HDDS-7566. Refactor RocksDBCheckpointDiffer tests.
Moved CompactionDag specific tests from TestRocksDBCheckpointDiffer to a new TestCompactionDag class.
This makes the tests more organized and easier to maintain.
